+++ lwb-music/package.json
@@ -16,12 +16,15 @@
         "@testing-library/react": "^11.1.0",
         "@testing-library/user-event": "^12.1.10",
         "cors": "^2.8.5",
+        "crypto-browserify": "^3.12.1",
         "dotenv": "^16.6.1",
         "emailjs": "^4.0.0",
         "emailjs-com": "^3.2.0",
         "express": "^4.18.0",
         "framer-motion": "^4.1.17",
         "nodemailer": "^6.7.3",
+        "os-browserify": "^0.3.0",
+        "path-browserify": "^1.0.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-icons": "^4.3.1",
