+++ lwb-music/src/pages/CheckoutPage.js
@@ -1,58 +1,40 @@
-// import React from 'react';
-// import { Elements } from '@stripe/react-stripe-js';
-// import CheckoutForm from '../components/CheckoutForm'; // Your custom form component
-// import { loadStripe } from '@stripe/stripe-js';
-
-// const stripePromise = loadStripe(process.env.STRIPE_PUBLISHABLE_KEY);
-
-// const CheckoutPage = () => {
-//   return (
-//    <Elements stripe={stripePromise}>
-//      <CheckoutForm />
-//     </Elements>
-//   );
-// };
-
-// export default CheckoutPage;
-
-//new comment here
-
-import * as React from 'react';
-import { useCallback } from 'react';
-import {loadStripe} from '@stripe/stripe-js';
-import {
-  EmbeddedCheckoutProvider,
-  EmbeddedCheckout
-} from '@stripe/react-stripe-js';
-require('dotenv').config();
-// Make sure to call `loadStripe` outside of a component’s render to avoid
-// recreating the `Stripe` object on every render.
+import React from 'react';
+import { Elements } from '@stripe/react-stripe-js';
+import { loadStripe } from '@stripe/stripe-js';
 
 const stripePromise = loadStripe(process.env.STRIPE_PUBLISHABLE_KEY);
 
 const CheckoutPage = () => {
-  const fetchClientSecret = useCallback(() => {
-    // Create a Checkout Session
-    return fetch("/create-checkout-session", {
-      method: "POST",
-    })
-      .then((res) => res.json())
-      .then((data) => data.clientSecret);
-  }, []);
-
-  const options = {fetchClientSecret};
-
-
   return (
-    <div id="checkout">
-      <EmbeddedCheckoutProvider
-        stripe={stripePromise}
-        options={options}
-      >
-        <EmbeddedCheckout />
-      </EmbeddedCheckoutProvider>
-    </div>
-  )
-}
+   <Elements stripe={stripePromise}>
+        <section>
+        <div className="product">
+          <img
+            src="https://i.imgur.com/EHyR2nP.png"
+            alt="The cover of Stubborn Attachments"
+          />
+          <div className="description">
+          <h3>Stubborn Attachments</h3>
+          <h5>$20.00</h5>
+          </div>
+        </div>
+        <form action="/checkout" method="POST">
+          <button type="submit">
+            Checkout
+          </button>
+        </form>
+      </section>
+    </Elements>
+  );
+};
 
 export default CheckoutPage;
+
+//new comment here
+
+// import React, { useState, useEffect } from "react";
+
+// const CheckoutForm = () => (
+  
+// );
+
